refactor : Jdbc 설정 한 클래스에서 관리로 수정
