+++ src/test/java/com/programmers/voucher/repository/voucher/VoucherRepositoryTest.java
@@ -1,5 +1,6 @@
 package com.programmers.voucher.repository.voucher;
 
+import com.programmers.voucher.JdbcConfig;
 import com.programmers.voucher.controller.dto.CustomerDto;
 import com.programmers.voucher.model.customer.Customer;
 import com.programmers.voucher.model.voucher.FixedAmountVoucher;
@@ -7,81 +8,20 @@ import com.programmers.voucher.model.voucher.PercentDiscountVoucher;
 import com.programmers.voucher.model.voucher.Voucher;
 import com.programmers.voucher.model.voucher.VoucherType;
 import com.programmers.voucher.repository.customer.CustomerRepository;
-import com.wix.mysql.EmbeddedMysql;
-import com.wix.mysql.ScriptResolver;
-import com.wix.mysql.config.Charset;
-import com.wix.mysql.config.MysqldConfig;
-import com.wix.mysql.distribution.Version;
-import com.zaxxer.hikari.HikariDataSource;
-import org.junit.jupiter.api.*;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.jdbc.DataSourceBuilder;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
-import org.springframework.test.context.junit.jupiter.SpringJUnitConfig;
 
-import javax.sql.DataSource;
 import java.util.List;
 import java.util.UUID;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-@SpringJUnitConfig
-@TestInstance(TestInstance.Lifecycle.PER_CLASS)
-class VoucherRepositoryTest {
+class VoucherRepositoryTest extends JdbcConfig {
     @Autowired
     private VoucherRepository voucherRepository;
     @Autowired
     private CustomerRepository customerRepository;
-    @Autowired
-    DataSource dataSource;
-    private static EmbeddedMysql embeddedMysql;
-
-    @Configuration
-    @ComponentScan(basePackages = {"com.programmers.voucher.repository"})
-    static class Config {
-
-        @Bean
-        public DataSource dataSource() {
-            return DataSourceBuilder.create()
-                    .url("jdbc:mysql://localhost:2215/test_voucher")
-                    .username("test")
-                    .password("test1234!")
-                    .type(HikariDataSource.class)
-                    .build();
-        }
-
-        @Bean
-        public JdbcTemplate jdbcTemplate(DataSource dataSource) {
-            return new JdbcTemplate(dataSource);
-        }
-
-        @Bean
-        public NamedParameterJdbcTemplate namedParameterJdbcTemplate(JdbcTemplate jdbcTemplate) {
-            return new NamedParameterJdbcTemplate(jdbcTemplate);
-        }
-    }
-
-    @BeforeAll
-    static void setup() {
-        MysqldConfig config = MysqldConfig.aMysqldConfig(Version.v8_0_17)
-                .withCharset(Charset.UTF8)
-                .withPort(2215)
-                .withUser("test", "test1234!")
-                .withTimeZone("Asia/Seoul")
-                .build();
-        embeddedMysql = EmbeddedMysql.anEmbeddedMysql(config)
-                .addSchema("test_voucher", ScriptResolver.classPathScript("schema.sql"))
-                .start();
-    }
-
-    @AfterAll
-    static void cleanup() {
-        embeddedMysql.stop();
-    }
 
     private Customer insertSingleCustomerData() {
         String email = "taehee@gmail.com";
